+++ app/src/main/java/com/example/quanlychitieu/domain/model/ChiTieuModel.kt
@@ -5,7 +5,7 @@ data class ChiTieuModel(
     val id_nguoidung: Int,
     val id_khoanchi: Int,
     val id_taikhoan: Int,
-    val so_tien: Int,
+    val so_tien: Long,
     val ngay_tao: String,
     val ghi_chu: String,
 )
