+++ app/src/main/java/com/example/quanlychitieu/domain/model/TaiKhoanModel.kt
@@ -4,7 +4,7 @@ data class TaiKhoanModel(
     val id: Int,
     val id_nguoidung: Int,
     val ten_taikhoan: String,
-    val so_du: Int,
+    val so_du: Long,
     val loai_taikhoan: Int,
     val mo_ta: String,
 )
